update the ami build process to use the previous ami
this vastly speeds up the packer ami build, by basing the build off a recent
straitjacket ami rather than the base ubuntu ami. that way we don't have to
re-build all the docker images from scratch.

+++ deploy/setup.sh
@@ -6,21 +6,24 @@ INSTALL_DIR=/home/ubuntu/straitjacket
 # write a blank config file, to be replaced by cloud-init
 touch /etc/straitjacket.env
 
+rm -rf $INSTALL_DIR > /dev/null
 mkdir $INSTALL_DIR
 tar xvof /home/ubuntu/straitjacket.tar -C $INSTALL_DIR
 chown -R ubuntu:ubuntu $INSTALL_DIR
 
 apt-get update
 apt-get install -y btrfs-tools
-mkfs.btrfs /dev/xvdb
-mkdir /var/lib/docker
-mount /dev/xvdb /var/lib/docker
-echo /dev/xvdb /var/lib/docker btrfs defaults 0 0 >> /etc/fstab
-wget -qO- https://get.docker.com/ | sh
-usermod -aG docker ubuntu
-systemctl enable docker
+if ! mount | grep -q /var/lib/docker; then
+  mkfs.btrfs /dev/xvdb
+  mkdir /var/lib/docker
+  mount /dev/xvdb /var/lib/docker
+  echo /dev/xvdb /var/lib/docker btrfs defaults 0 0 >> /etc/fstab
+  wget -qO- https://get.docker.com/ | sh
+  usermod -aG docker ubuntu
+  systemctl enable docker
+fi
 
-cd /home/ubuntu/straitjacket
+cd $INSTALL_DIR
 
 ./straitjacket-setup.sh
